INSERTING EVENTS!

+++ Aphro-WebForms/Models/Building.cs
@@ -7,7 +7,7 @@ namespace Aphro_WebForms.Models
 {
     public class Building
     {
-        public long building_key { get; set; }
+        public int building_key { get; set; }
         public string description { get; set; }
     }
 }
